Refactored subtest files to utils
