+++ tests/aswwu/behaviors/elections/candidate/candidate_utils.py
@@ -1,4 +1,4 @@
-import tests.aswwu.behaviors.elections.election.election_subtests as election_subtests
+import tests.aswwu.behaviors.elections.election.election_utils as election_utils
 import tests.aswwu.behaviors.elections.candidate.candidate_requests as candidate_requests
 import tests.aswwu.behaviors.elections.position.position_requests as position_requests
 import tests.aswwu.behaviors.auth.auth_requests as auth_requests
@@ -52,7 +52,7 @@ def create_default_candidate_params(session):
     :return election_id: generic election id
     :return position_ids: generic position ids
     """
-    election_id = election_subtests.assert_post_dynamic_election(session)['id']
+    election_id = election_utils.assert_post_dynamic_election(session)['id']
     position_resp_aswwu = position_requests.post_position(session,
                                                           position=position_data.ASWWU_DATA['position'],
                                                           election_type=position_data.ASWWU_DATA['election_type'],
